requirements.txt
aws-cdk-lib==2.170.0
constructs>=10.0.0,<11.0.0
aws-lambda-powertools==3.4.1
aws-xray-sdk==2.14.0
cdk-nag==2.35.14
boto3==1.35.99
botocore==1.35.99
certifi==2024.12.14 
cffi==1.17.1
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
PyYAML==6.0.2
requests==2.32.3
responses==0.25.6
urllib3==1.26.19
aws-psycopg2==1.3.8
numpy==2.2.2
pandas==2.2.3